requirement.txt
pyyaml
jiagu
numpy
pymysql
requests
tiktoken